improve: code structure
