+++ internal/benchmark/locomo/observation_store.go
@@ -1,0 +1,111 @@
+package locomo
+
+import (
+	"bufio"
+	"bytes"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+func readObservationRecords(path string) (map[string]ObservationRecord, error) {
+	file, err := os.Open(path)
+	if errors.Is(err, os.ErrNotExist) {
+		return map[string]ObservationRecord{}, nil
+	}
+	if err != nil {
+		return nil, fmt.Errorf("open benchmark observations: %w", err)
+	}
+	defer file.Close()
+	result := make(map[string]ObservationRecord)
+	scanner := bufio.NewScanner(file)
+	scanner.Buffer(make([]byte, 64<<10), maxObservationBytes)
+	for line := 1; scanner.Scan(); line++ {
+		if len(strings.TrimSpace(scanner.Text())) == 0 {
+			continue
+		}
+		var value ObservationRecord
+		if err := json.Unmarshal(scanner.Bytes(), &value); err != nil {
+			return nil, fmt.Errorf("decode benchmark observation line %d: %w", line, err)
+		}
+		if value.QuestionID == "" {
+			return nil, fmt.Errorf("benchmark observation line %d has no question_id", line)
+		}
+		result[value.QuestionID] = value
+	}
+	if err := scanner.Err(); err != nil {
+		return nil, fmt.Errorf("read benchmark observations: %w", err)
+	}
+	return result, nil
+}
+
+func appendJSONLine(path string, value any) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
+		return err
+	}
+	encoded, err := compactJSON(value)
+	if err != nil {
+		return err
+	}
+	file, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o644)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	if _, err := file.Write(append(encoded, '\n')); err != nil {
+		return err
+	}
+	return file.Sync()
+}
+
+func compactJSON(value any) ([]byte, error) {
+	var output bytes.Buffer
+	encoder := json.NewEncoder(&output)
+	encoder.SetEscapeHTML(false)
+	if err := encoder.Encode(value); err != nil {
+		return nil, err
+	}
+	return bytes.TrimSuffix(output.Bytes(), []byte{'\n'}), nil
+}
+
+func writeJSON(path string, value any) error {
+	var output bytes.Buffer
+	encoder := json.NewEncoder(&output)
+	encoder.SetEscapeHTML(false)
+	encoder.SetIndent("", "  ")
+	if err := encoder.Encode(value); err != nil {
+		return err
+	}
+	return writeFileAtomic(path, output.Bytes())
+}
+
+func writeFileAtomic(path string, contents []byte) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
+		return err
+	}
+	temporary, err := os.CreateTemp(filepath.Dir(path), ".powercontext-locomo-*")
+	if err != nil {
+		return err
+	}
+	temporaryPath := temporary.Name()
+	defer os.Remove(temporaryPath)
+	if err := temporary.Chmod(0o644); err != nil {
+		_ = temporary.Close()
+		return err
+	}
+	if _, err := temporary.Write(contents); err != nil {
+		_ = temporary.Close()
+		return err
+	}
+	if err := temporary.Sync(); err != nil {
+		_ = temporary.Close()
+		return err
+	}
+	if err := temporary.Close(); err != nil {
+		return err
+	}
+	return os.Rename(temporaryPath, path)
+}
