--- internal/sqlstore/sqlite/doc.go
@@ -1,2 +1,0 @@
-// Package sqlite implements the SQLite relational profile and search indexes.
-package sqlite
