--- internal/testkit/doc.go
@@ -1,2 +1,0 @@
-// Package testkit contains deterministic fixtures shared by internal tests.
-package testkit
